Added Pandas and created the Matcher class:
1. Created the matcher class.

The matcher class will be what can be
imported and used to match cases and controls. The class has three
attributes: cases, controls, and match_ratio. Cases/controls are both dataframes that have the
ids and information for cases/controls, respectively. The information
will be used to match cases and controls. The match_ratio attribute is
an integer indicating how many controls should be matched to a case.

The class exposes a method called match that the user can use after
initializing the class. this method takes a mating function that either
the user can define or the default Matcher.sex_age_match function. These
functions need to accept the three keyword arguments: cases, controls,
and ratio

1. Added pandas as a dependency.

The matcher class takes cases and controls as a pandas dataframe